asistencias/requirements.txt

Django==3.1

# database
psycopg2-binary==2.8.5

djangorestframework==3.13.1
django-filter==21.1
